fix: remove cast

+++ solidity/test/unit/PrivateERC20ResolutionModule.t.sol
@@ -63,7 +63,7 @@ contract PrivateERC20ResolutionModule_UnitTest is Test {
   // Mock EOA disputer
   address public disputer;
 
-  event CommitingPhaseStarted(uint128 _startTime, bytes32 _disputeId);
+  event CommitingPhaseStarted(uint256 _startTime, bytes32 _disputeId);
   event VoteCommited(address _voter, bytes32 _disputeId, bytes32 _commitment);
   event VoteRevealed(address _voter, bytes32 _disputeId, uint256 _numberOfVotes);
   event DisputeResolved(bytes32 _disputeId, IOracle.DisputeStatus _status);
@@ -111,7 +111,7 @@ contract PrivateERC20ResolutionModule_UnitTest is Test {
 
     // Check: emits CommitingPhaseStarted event?
     vm.expectEmit(true, true, true, true);
-    emit CommitingPhaseStarted(uint128(block.timestamp), _disputeId);
+    emit CommitingPhaseStarted(block.timestamp, _disputeId);
 
     vm.prank(address(oracle));
     module.startResolution(_disputeId);
